Struct rb::RbList

pub struct RbList<O: Op> { /* private fields */ }

A list based on a red-black tree.

Implementations

impl<O: Op> RbList<O>

pub fn new() -> Self

Constructs a new, empty list.

pub fn is_empty(&self) -> bool

Returns the length of the list.

pub fn len(&self) -> usize

Returns the length of the list.

pub fn front(&self) -> Option<&O::Value>

Provides a reference to the front element, or None if the deque is empty.

pub fn back(&self) -> Option<&O::Value>

Returns the last element of the list.

pub fn front_mut(&mut self) -> Option<&mut O::Value>

Returns the first element of the list.

pub fn back_mut(&mut self) -> Option<&mut O::Value>

Returns the last element of the list.

pub fn push_front(&mut self, value: O::Value)

Prepends an element to self.

pub fn push_back(&mut self, value: O::Value)

Appends an element to self.

pub fn entry(&mut self, index: usize) -> Entry<'_, O>

Returns a deligated entry, which is a mutable reference to the element at the given index.

pub fn pop_front(&mut self) -> Option<O::Value>

Removes the first element and returns it, or None if the deque is empty.

pub fn pop_back(&mut self) -> Option<O::Value>

Removes the last element and returns it, or None if the deque is empty.

pub fn insert(&mut self, index: usize, value: O::Value)

Inserts an element at the given index.

pub fn remove(&mut self, index: usize) -> O::Value

Removes the element at the given index.

pub fn append(&mut self, other: &mut Self)

Moves all elements from other into self, leaving other empty.

pub fn split_off(&mut self, index: usize) -> Self

Splits self into two at the given index.

pub fn partition_point<P>(&self, pred: P) -> usizewhere P: FnMut(&O::Value) -> bool,

Returns the index of the partition point according to the given predicate (the index of the first element of the second partition).

pub fn binary_search<F>(&self, f: F) -> Result<usize, usize>where F: FnMut(&O::Value) -> Ordering,

Binary searches self for a given element

pub fn iter(&self) -> Range<'_, O>

Iterates over the list.

pub fn range(&self, range: impl RangeBounds<usize>) -> Range<'_, O>

Returns an iterator over the given range.

pub fn fold<R: RangeBounds<usize>>(&self, range: R) -> Option<O::Acc>

Returns the product of all the elements within the range.