Trait rb::Op

pub trait Op {
    type Value;
    type Acc;
    type Action;

    // Required methods
    fn to_acc(value: &Self::Value) -> Self::Acc;
    fn mul(lhs: &Self::Acc, rhs: &Self::Acc) -> Self::Acc;
    fn apply(value: &mut Self::Value, lazy: &Self::Action);
    fn apply_acc(acc: &mut Self::Acc, lazy: &Self::Action);
    fn compose(lhs: &Self::Action, rhs: &Self::Action) -> Self::Action;
    fn identity_action() -> Self::Action;
}

The trait for specifying the operation of a red-black tree.

Notation

$xy$ denotes the multiplication of two accumulated values.

$a \circ b$ denotes the composition of two lazy actions.

$x \cdot a$ denotes the application of a lazy action to an accumulated value.

Laws

For RbList:

mul is associative. ($(xy)z = x(yz)$)

compose is associative. ($(a \circ b) \circ c = a \circ (b \circ c)$)

apply is an action of a semigroup on a semigroup. ($(xy) \cdot a = (x \cdot a)(y \cdot a)$, $x \cdot (a \circ b) = (x \cdot a) \cdot b$)

Furthermore, for RbReversibleList:

mul is commutative. ($xy = yx$)

Required Associated Types

type Value

The type of the value stored in a node.

type Acc

The type of the accumulated value of a node.

type Action

The type of the lazy action of a node.

Required Methods

fn to_acc(value: &Self::Value) -> Self::Acc

Convert a value to an accumulated value.

fn mul(lhs: &Self::Acc, rhs: &Self::Acc) -> Self::Acc

Multiply two accumulated values.

fn apply(value: &mut Self::Value, lazy: &Self::Action)

Apply a lazy action to a value.

fn apply_acc(acc: &mut Self::Acc, lazy: &Self::Action)

Apply a lazy action to an accumulated value.

fn compose(lhs: &Self::Action, rhs: &Self::Action) -> Self::Action

Compose two lazy actions.

fn identity_action() -> Self::Action

The identity of an accumulated value.